Fix: Registry service may raise exception if `.register()` is used in constructor method of component
This happens because everytime we create a new instance of the
component, it's constructor method will be called. And RegistryService
forbids us to override registration. For example: If you logout and try to
login again, components will be recreated and error will be thrown on
attempt to override existing config.

We need to use a place where whatever code we put will run just once,
and Module constructor seems like a good place to do so. It's an
equivalent of `module.run()` in AngularJS.

I learned about this from this Github comment[1] by Miško Hevery,
Father of Angular.

[1] https://github.com/angular/angular/issues/1530#issuecomment-253406502